+++ lib/daru/data_tables.rb
@@ -1,0 +1,4 @@
+require 'daru/data_tables/version'
+require 'daru/data_tables/data_table'
+require 'daru/data_tables/display/iruby_notebook'
+require 'daru/data_tables/display/display'
