+++ lib/daru/data_tables/templates/init.inline.css.erb
@@ -1,7 +1,7 @@
 <%
 
 dependent_css.each{ |dep_css|
-  contents = IO.read(File.join(css_dir,dep_css), mode: 'r:UTF-8')
+  contents = File.read(File.join(css_dir,dep_css), mode: 'r:UTF-8')
 %>
    /* BEGIN <%= dep_css %> */
