+++ lib/daru/data_tables/templates/table_script.erb
