+++ lib/daru/data_tables/version.rb
@@ -1,0 +1,5 @@
+module Daru
+  module DataTables
+    VERSION = "0.1.0"
+  end
+end
