--- lib/data_tables.rb
@@ -1,5 +1,0 @@
-require "data_tables/version"
-require "data_tables/data_tables"
-require_relative "data_tables/display/iruby_notebook"
-require_relative "data_tables/display/display"
-
