--- lib/data_tables/version.rb
@@ -1,3 +1,0 @@
-module DataTables
-  VERSION = "0.1.0"
-end
