+++ spec/dummy_rails/app/views/layouts/application.html.erb
@@ -6,7 +6,7 @@
 
     <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
-    <%=raw DataTables.init_script %>
+    <%=raw Daru::DataTables.init_script %>
   </head>
 
   <body>
